+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+## 3.1.3 (2021-01-25)
+
+- Fix the fetching of the version of the SDK (#1169)
 - Add the `$customSamplingContext` argument to `Hub::startTransaction()` and `HubAdapter::startTransaction()` to fix deprecations thrown in Symfony (#1176)
 
 ## 3.1.2 (2021-01-08)